Refactor to split methods
